LPS-97253 Move for consistency
